requirements.txt
# Core Dependencies
langchain
langchain-community
langchain-openai
langchain-text-splitters
langchain-core
python-dotenv

# AI/ML
openai
tiktoken

# Vector Store
chromadb
supabase

# Discord
discord.py

# PDF Processing
pypdf
PyPDF2

# Data Validation & Settings
pydantic>=2.0.0
pydantic-settings>=2.0.0

# HTTP
httpx

# Documentation
mkdocs
mkdocs-material

# Testing (Development)
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0

# Code Quality (Development)
mypy>=1.0.0
ruff>=0.1.0
black>=23.0.0

# Type Stubs
types-requests
